// distilled-coinbase/src/operations/getPolicyById.ts
export const GetPolicyByIdInput = Schema.Struct({
  policyId: Schema.String.pipe(T.PathParam()),
}).pipe(T.Http({ method: "GET", path: "/v2/policy-engine/policies/{policyId}" }));
export type GetPolicyByIdInput = typeof GetPolicyByIdInput.Type;
// ...
export type GetPolicyByIdOutput = typeof GetPolicyByIdOutput.Type;

// The operation
/**
 * Get a policy by ID
 *
 * Get a policy by its ID.
 *
 * @param policyId - The ID of the policy to get.
 */
export const getPolicyById = /*@__PURE__*/ /*#__PURE__*/ API.make(() => ({
  inputSchema: GetPolicyByIdInput,
  outputSchema: GetPolicyByIdOutput,
}));